Separate out map center btn
